Weird Flecks, But OK (2019 NC NA Regional A)
This stupid thing. Oh my lord. Uses welzl's algorithm.
This could've been exhaustive but honestly, it's fine as is.

== NOTE ==
Does not work to simply take the max distance of x/y coords (or equiv.)
since the smallest circle could be determined by three points instead of two.

This is what was keeping my third test case from passing, but after that
it was free. Minus the segfaults n such, of course.